requirements.txt
Django==4.2.7
djangorestframework==3.14.0
lxml==4.9.3
networkx==3.2.1
ocpa==1.3.3
ocpa.egg==info